+++ autoextract_poet/exceptions.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+
+from autoextract_poet.query import Query
+
+
+@dataclass
+class QueryLevelError(Exception):
+
+    query: Query
+    msg: str
